Create README.md
This survey is undertaken to find out if students in general at I.U.P prefer living on campus or off campus. This helps us to understand if the students feel secure on campus and to get to know their opinion about the on-campus facilities, the diversity of cultures on campus, if living on campus is worth their money, and if they have good access to the numerous on campus resources, such as the faculty, library, academic advisors, computer labs, dining halls, and healthcare. It also helps us to analyze the level of responsibility they have on campus and off campus and their opinion on sharing option (rooms and washrooms).